ksum/requirements.txt
streamlit>=1.31.0
openai>=1.12.0
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
elevenlabs>=0.2.27
moviepy>=1.0.3
python-magic>=0.4.27
numpy>=1.24.0
pydub>=0.25.1 